week1/requirements.txt
sqlalchemy
pandas
python-decouple
psycopg2-binary